+++ AlgorithmComparison/src/main/java/algorithms/algorithmcomparison/runtimetest/AlgorithmRuntimeTester.java
@@ -14,7 +14,7 @@ public class AlgorithmRuntimeTester {
 	
 	public RuntimeStatistics run(ArrayList<int[]> arrays) {
 		long[] runtimeResults = new long[NUM_OF_ITERATIONS];
-		int[] keyCompResults = new int[NUM_OF_ITERATIONS];
+		long[] keyCompResults = new long[NUM_OF_ITERATIONS];
 		
 		for(int i = 0; i < NUM_OF_ITERATIONS; i++){
 			algorithm.reset();
